Add a Tabs landing page
The tabs landing page shows the tabs that are ready to be debugged.
When a tab is selected the child id is appended to the url hash so that
future refreshs will maintain that selection.

Changes:

There's a new Toolbox component, Tabs component, action, and reducer.
At some point it would be nice to pull these pieces out of the debugger.

The protocol logic was moved into a thread-client module, which exposes
connectToClient and connectToTab methods. This is probably the most
important thing to review.