The refactor cleans up many of the queries in two important ways:
1) composition
2) return types

Composition - there are a few queries (selectors) that access state properties directly.

Return types - queries now return the state's immutable js types. This extends the immutable js boundary significantly. It is now reasonable to expect a component propType to be an immutable type and to expect immutable types to be passed into actions. The new immutable js boundary is the client and reducers, which means that both the client and reducers expect plain JS objects. It is probably reasonable to have reducers expect immutable.js objects soon though.