Update Configuration
Currently we use an environment file for configuration.
This is fine, but there are a couple shortcomings

1) the environment file is read directly
2) the environment file is for your own local changes, which is not the
same as shared development features
3) it would not be obvious how to add production flags. This PR adds a
development.json file, but it would be easy to add a production.php file
and read from it when in production.
4) lastly, there aren't any nice helpers. This PR adds an isEnabled
feature, but ultimately it would be nice to improve it to support path
notation and maybe other functionaality as well.