Add breakpoint styling
This sets the line number font to white and moves the breakpoint below
the linenumber.

I couldn't find where Firefox did that work, so I thought I'd share the
PR incase there's a better way of doing it. One thing I noticed for
example, is if the breakpoint appears ahead in the DOM structure, then
it doesn't need any z-index treatment.