reduce the repetition of frame location
This is just a hack attempt at only listing the frame location when its
unique to the frame before it.

Currently you can the same frame source location repeated over and
over.  While the line number may be unique and should be displayed we
should avoid repeating ourselves like this.