Browser Tab cleanup work (chrome part 3)
This is just a refactor PR, nothing new was added.

+ cleaned up the way tab ids were managed:
Previously, tab Ids were either firefox actors or chrome tab ids. This
created an issue when it came to persisting an ID in the url query.
Previously, we were keeping just the child/tab of the firefox actor, but
now that it can be either firefox or chrome, the ID was easier. So, to
make this work, the tab ID is now either the chrome tab id or the
firefox child or tab part of the actor.

You might be asking, "why not use the entire actor?" The reason is that
the tab actor includes a connection id which is incremented everytime
the debugger is refreshed, which is exactly what we're trying to avoid.
What about if the user changes the tab order in firefox and then
refreshes the debugger? This will probably cause the debugger to debug
a different tab. We should check though...

Reasons:

+ debugTab seems like something you would want in your redux history
+ it's nice for the one action to know to call the firefox or chrome
client debugTab functions
+ it's nice for it to first invoke the select tab action.