Replace karma w/ firefox driver
This PR replaces the complicated karma overhead with a simple firefox
driver approach that loads the mocha server page and reads the results.

We can always expand the output by reading the test results or setting
up a separate reporter, but for now this gives us guidance on whether
the unit tests are running successfully.

Leaving Karma:

* we've already concluded that we are not looking to test specific
   browser apis
* karma introduces a lot of overhead, which is unecessary for what we
  are trying to do.