move network timeout code to connect() (#624)
Catch errors and reject the promise in chrome client
Remove the timeout function from the firefox client and move it to the
shared/client/main DebuggerClient.connect function
Use a similar Promise.race() style for handling the timeout in both
networkRequest and DebuggerClient.connect
In the DebuggerClient I used the new Promise() style of promises but
the rest of the code uses deferred.  I wanted to keep a similar style
to the rest of the code but it just feels awkward here.  I can rework
it to use deferred if that’s needed.