Rename l10n keys
We previously bumped the key name for translation reasions, but
it turns out that if the old translation is okay then it's alright to
change the value.

This is helpful for us because when we changed the key name we actually
broke the old debugger. If we do want to change an L10N string we should
copy it and create a new one so we don't break the old debugger.