Remove devtools.debugger.client-source-maps-enabled pref
This removes the devtools.debugger.client-source-maps-enabled pref, in
favor of the pref that will be used across all tools.

I've copied over the relevant from devtools.js; the next time the
debugger lands, we will want to remove that definition from M-C.

I'll handle other aspects of the unification in M-C; namely, updating
the options panel.