[WIP] Add an openLink action mapped to toolbox openUILinkIn.
This is working in launchpad and in toolbox.
I am not sure that this should be in an action though,
but I didn't want to change too much things to pass it
as a prop to the App components and pipe it down.

One thing to have in mind is that we will also need the
toolbox integration for the open-in-inspector button on
DomElement reps. So maybe all the toolbox integration
can be done in the actions/toolbox.js file.

I plan to add a mochitest to make sure we don't regress
on this in the different components.