scopes.js getScopes: Fix <return> and <exception> pseudo-vars.
As of Bug 1370923, frames in "paused" packets from the server no longer include
an environment form. This broke getScopes, which assumed it could recognize the
scope of the frame that is currently returning by comparing the scope's actor
name with that given in the "paused" packet's frame.

Since only one frame at a time will ever have a frameFinished property, it's
good enough to simply watch for the top scope of the selected frame, and append
the values whenever the frame has a frameFinished property.