Fix ObjectInspector indent issue.
In #5269, we introduced an override for non-expandable nodes on the Tree component,
since we don't use the provided arrow in the project.
But, by doing that, we introduced a regression in the ObjectInspector indent,
since it also uses the Tree component.

Here we add more specificity to the overriding rule so we don't target ObjectInspector anymore.