Remove references to `listTabs` API
`listTabs` is bad for performance and should be avoided where possible.  In most
cases of interest to the debugger, a `target` is already prepared by the toolbox
with all data it would need.

The debugger doesn't actually seem to use `listTabs` currently, so we should
remove these types as well to avoid the suggestion that it's a good idea to use.