Make stacktrace copy/paste more useful.
At the moment when one try to copy and paste the stacktrace
of an error object, everything gets mangled in the same line,
making it very hard to read.
This patch adds a tab character before each function name as
well as a new line after each frame so what the user gets in
the clipboard looks more like what they see on the output.