Fix bug in Error stacktrace generation when stacktrace is a longString.
In order to build the stacktrace we parse the string we
receive from the server. In some cases, the stacktrace
can be very long and a longString is sent instead of a
string. This means that the property that we check might
have a cut-off location at the end.
Due to a placement error in an if, it could happen that
this last cut-off element would throw.
The fix is made by correctly guarding the property we
are checking.
We take this as an opportunity to change how we handle
the last frame (we don't include it in the stack array
we build, instead of putting it and then removing it).
A snapshot test is added to make sure everything is
working as expected.