Fix expanding longString when the property is a getter.
When we have a getter value, it is stored in a different property
in the node. Which means that if the getter returns a longString,
we couldn't assign the fullText to the non-getter property (and as
a result, an exception was thrown).
This is not a common bug since we only automatically evaluate safe
getters, i.e. native property getters that we know are side-effect
free.

A test is added to make sure we handle this as expected.