Roll back Tree component onFocus change.
The check that was made there prevented the first
item of the Tree to be selected if the click happened
inside it. Without this check, the first item is automatically
selected and scrolled to, which is causing some weirdness in
the console (probably in the debugger as well).
The fix was made in mozilla-central directly in the reps bundle
as well.