Remove inToolbox function in production.
When we're generating bundles for mozilla-central,
there's no reason we should keep the inToolbox
functions as the answer is always true.
We remove the function declaration and replace
the call to the function by true.
This fixes an issue we have while trying to start
the parser worker from the toolbox (window is undefined).