Handle Frame overflow (function and location) in CSS.
The Frames component has limited space available,
and it might happen that we need to truncate some
information.
Previouysly, this was done by manually setting a
length limit on the string we were displaying.
This is a bit unfortunate since CSS has some useful
properties we can use to show whatever fits on the
screen.
So now, in Frames, we don't truncate strings anymore,
and, in CSS, we set the text-overflow we want.
Furthermore, we make the Frame element wrap, so when
the panel gets smaller, the location is put on a
new line instead of shrinking (and then shrinks if
it needs to).
In order to have such feature, we needed to have the
location element to not have display: flex. This was
not hard to fix since flex was only used to place an
optional icon after the framework name, which can be
do with inline display.