Handle anonymous functions more consistently.
Anonymous function name was not displayed
at all in the callstack panel, neither it was
in PreviewFunction. In the scopes panel, it was
displayed in another way (`(anonymous)`) than
in the toolbox, where we use `<anonymous>`.
With this patch, we make sure to have anonymous
functions displayed in a consistent way in the debugger,
and matching other panels in the toolbox.

In order to do that, the anonymous item in the properties
file is updated to match the existing ones in mozilla-central.