Fix Tree arrow images.
We used to render the arrows using an empty img
with a CSS mask. Due to some platform changes,
we now see a bit of the broken image asset,
which does not look great.

In order to fix that, we use buttons to render
the arrow, with a background-image, and context-fill
to style it.