Invoke getter in the root object.
To invoke getter, we were retrieving the parent node
so we could access the getter. This works fine unless
the getter is itself somewhere in the prototype chain
and refers to a property set by an upper level.
This patch changes how we retrieve the context to
invoke the getter in to be the top-most, non-prototype
grip, which should cover this case.