Fix CallStack Group with anonymous function.
formatDisplayName was called without the l10n
object, which caused an exception if the last
frame of a stack was anonymous (because we then
use l10n).
The fix is simple, pass the l10n object to
formatDisplayName.
A test is added to ensure we cover this case.
The test do fails without the fix in Group.js