Improve the expand/collapse all behaviour of the Sources tree
First, remove the expansion code in the SourcesTreeItem onClick
event as the expansion is already handled in the onClick event
of the tree component.

Next, pass the altKey status from the tree onClick event to the
ManagedTree onExpand/onCollapse functions so that we know whether
to expand/collapse the children.