Fix unhelpful error in top-level await malformed expression.
In mapExpression, if we fail to get an AST
from the expression this means the expression
has a syntax error in it.
In order to get a meaningful error when evaluating
top-level await expression, we still need to
wrap the expression in an async iife.
This required some changes in mapExpression in order
to not bail completely if parsing the expression
threw.
We also take this as an opportunity to only parse
the expression once for the all the transforms.