Render href attribute in String rep URLs.
Fixes #6108.

The attribute was removed because in some places
the link could be opened directly in the panel's
frame, which isn't good.
But this is breaking JSON Viewer links, and also
prevent some context menu entries to work in the
console (open link, copy link, …).

This patch adds the href attribute back, but only
if the reps are meant to be displayed in content
page, or if an openLink function is provided.